+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Types of changes are:
 
 ## Unreleased
 
+### Fixed
+- `Select`: Fix responsiveness and added clearable icon to clear option
+
 ## 2.0.14 - 2025-10-07
 ### Added
 - New system icons (cloud-upload)